requirements.txt
flwr
shap
lime
scikit-surprise
numpy<2.0
recommenders
